notes from video 19

+++ cajon/platzi/avanzado_de_php/portafolio/app/Services/JobService.php
@@ -10,6 +10,10 @@ class JobService
 	public function deleteJob($id)
 	{
 		$job = Job::find($id);
+		if (!$job)
+		{
+		    throw new \Exception( 'Job not found');
+        }
 		$job->delete();
 	}
 }
